pre rendering issues resolved
